chore(...) : 핫 키워드, 결과 viewModel 분기 및 패키지이동
