+++ app/src/main/AndroidManifest.xml
@@ -55,7 +55,7 @@
         <activity android:name=".presentation.home.setting.SettingNoticeActivity" />
         <activity android:name=".presentation.home.study.StudyCategoryActivity" />
         <activity android:name=".presentation.home.study.StudyCreateActivity" />
-        <activity android:name=".presentation.home.study.SearchStudyActivity" />
+        <activity android:name=".presentation.home.study.search.SearchStudyActivity" />
         <activity
             android:name=".presentation.place.SearchPlaceDetailActivity"
             android:windowSoftInputMode="adjustResize" />
