+++ app/src/main/java/com/iron/espresso/presentation/home/study/StudyFragment.kt
@@ -12,6 +12,7 @@ import com.iron.espresso.presentation.StudyCategoryItem
 import com.iron.espresso.presentation.home.study.adapter.CategoryAdapter
 import com.iron.espresso.presentation.home.study.adapter.viewholder.StudyCategoryAdapterListener
 import com.iron.espresso.presentation.home.study.list.StudyListActivity
+import com.iron.espresso.presentation.home.study.search.SearchStudyActivity
 
 class StudyFragment :
     BaseFragment<FragmentStudyBinding>(R.layout.fragment_study),
