+++ app/src/main/java/com/iron/espresso/presentation/home/study/search/HotKeywordViewModel.kt
@@ -1,0 +1,35 @@
+package com.iron.espresso.presentation.home.study.search
+
+import androidx.hilt.lifecycle.ViewModelInject
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import com.iron.espresso.base.BaseViewModel
+import com.iron.espresso.model.api.StudyApi
+import com.iron.espresso.presentation.home.study.HotKeywordItem
+
+class HotKeywordViewModel @ViewModelInject constructor(private val studyApi: StudyApi) :
+    BaseViewModel() {
+
+    private val _hotKeywordList = MutableLiveData<List<HotKeywordItem>>()
+    val hotKeywordList: LiveData<List<HotKeywordItem>>
+        get() = _hotKeywordList
+
+    fun getHotKeywordList() {
+        val hotKeywordItemList = arrayListOf<HotKeywordItem>().apply {
+            add(HotKeywordItem("프로젝트"))
+            add(HotKeywordItem("Swift"))
+            add(HotKeywordItem("안드로이드fdsafdsafdsafdsafdsafdsa"))
+            add(HotKeywordItem("node.rkdcjfajdcjddl"))
+            add(HotKeywordItem("코드리뷰"))
+            add(HotKeywordItem("안드로이드"))
+            add(HotKeywordItem("node.js"))
+            add(HotKeywordItem("코드리뷰fdsafdsafdsafdsafdsafdsafd"))
+            add(HotKeywordItem("취업스터디"))
+            add(HotKeywordItem("취업스터디"))
+            add(HotKeywordItem("프로젝트fdsafdsafdsafdsafdsa"))
+            add(HotKeywordItem("Swift"))
+        }
+
+        _hotKeywordList.value = hotKeywordItemList
+    }
+}
