+++ app/src/main/java/com/iron/espresso/presentation/home/study/search/StudyResultViewModel.kt
@@ -1,4 +1,4 @@
-package com.iron.espresso.presentation.home.study
+package com.iron.espresso.presentation.home.study.search
 
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
@@ -13,16 +13,12 @@ import com.iron.espresso.ext.toErrorResponse
 import com.iron.espresso.model.api.StudyApi
 import retrofit2.HttpException
 
-class SearchStudyViewModel @ViewModelInject constructor(private val studyApi: StudyApi) :
+class StudyResultViewModel @ViewModelInject constructor(private val studyApi: StudyApi) :
     BaseViewModel() {
 
     private val _studyList = MutableLiveData<List<StudyItem>>()
     val studyList: LiveData<List<StudyItem>>
         get() = _studyList
-
-    private val _hotKeywordList = MutableLiveData<List<HotKeywordItem>>()
-    val hotKeywordList: LiveData<List<HotKeywordItem>>
-        get() = _hotKeywordList
 
     private val allList = mutableListOf<StudyItem>()
     private var pagingSize = -1
@@ -115,24 +111,5 @@ class SearchStudyViewModel @ViewModelInject constructor(private val studyApi: St
                     }
                 })
         }
-    }
-
-    fun getHotKeywordList() {
-        val hotKeywordItemList = arrayListOf<HotKeywordItem>().apply {
-            add(HotKeywordItem("프로젝트"))
-            add(HotKeywordItem("Swift"))
-            add(HotKeywordItem("안드로이드fdsafdsafdsafdsafdsafdsa"))
-            add(HotKeywordItem("node.rkdcjfajdcjddl"))
-            add(HotKeywordItem("코드리뷰"))
-            add(HotKeywordItem("안드로이드"))
-            add(HotKeywordItem("node.js"))
-            add(HotKeywordItem("코드리뷰fdsafdsafdsafdsafdsafdsafd"))
-            add(HotKeywordItem("취업스터디"))
-            add(HotKeywordItem("취업스터디"))
-            add(HotKeywordItem("프로젝트fdsafdsafdsafdsafdsa"))
-            add(HotKeywordItem("Swift"))
-        }
-
-        _hotKeywordList.value = hotKeywordItemList
     }
 }
