+++ app/src/main/res/layout/activity_search_study.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.home.study.SearchStudyActivity">
+        tools:context=".presentation.home.study.search.SearchStudyActivity">
 
         <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/appbar"
